chore: relocate color_default.py and also the default colors injection logic
