+++ flexmeasures/ui/utils/color_defaults.py
